requirements.txt
aioredis==1.3.1
argon2-cffi==19.1.0
asgiref==3.2.7
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.6.2
Automat==0.7.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
certifi==2019.6.16
cffi==1.12.3
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.5.0
decorator==4.0.10
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-braces==1.13.0
django-channels==0.7.0
django-cors-headers==3.1.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.0
-e git+https://github.com/django-extensions/django-extensions.git@4e9989efac51f562b6ec9827095cd4012d2b0ebf#egg=django_extensions
django-filter==2.2.0
django-mediumeditor==1.0.0
django-ratings==0.3.7
django-rest-knox==4.1.0
django-summernote==0.8.11.6
djangorestframework==3.10.1
djangorestframework-simplejwt==4.4.0
enum34==1.1.6
Faker==1.0.7
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython-genutils==0.1.0
Markdown==3.1.1
msgpack==0.6.2
oauthlib==3.0.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.0.0
pipenv==2018.11.26
prompt-toolkit==1.0.9
psycopg2==2.8.5
ptyprocess==0.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.19
pycryptodome==3.8.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pypiwin32==223
python-dateutil==2.8.0
python-decouple==3.3
python3-openid==3.1.0
pyttsx3==2.71
pytz==2019.1
pywin32==224
requests==2.22.0
requests-oauthlib==1.2.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
SpeechRecognition==3.8.1
sqlparse==0.3.0
static3==0.7.0
text-unidecode==1.2
toolz==0.10.0
twilio==6.43.0
Twisted==20.3.0
txaio==18.8.1
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
wikipedia==1.4.0
zope.interface==4.6.0
